requirements.txt
fastapi
uvicorn
redis
requests
pydantic
playwright
opencv-python
python-dotenv
google-generativeai
websockets
aiohttp

